Deleted duplicated class SCMExtensionParser from SCM package (duplicate of class SVNExtensionParser). Renamed SVNExtensionParser to SVNOffineLogParser.
